Clean up packages: add sim, model, and alg
