+++ src/main/java/com.ai/Policy.java
@@ -1,5 +1,8 @@
 package com.ai;
 
+import com.ai.model.Action;
+import com.ai.model.State;
+
 public interface Policy {
     public Action getAction(State state);
 }
