+++ src/main/java/com.ai/model/Position.java
@@ -1,0 +1,41 @@
+package com.ai.model;
+
+import java.util.Objects;
+
+public class Position {
+    private final int x;
+    private final int y;
+
+    public Position(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(x, y);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if(!(o instanceof Position)) {
+            return false;
+        }
+        Position other = (Position)o;   
+        return this.x == other.x &&
+               this.y == other.y;
+    }
+
+    @Override
+    public String toString() {
+        return "(" + x + ", " + y + ")";
+    }
+}
