+++ src/main/java/com.ai/sim/ActionSimulator.java
@@ -1,4 +1,7 @@
-package com.ai;
+package com.ai.sim;
+
+import com.ai.model.Action;
+import com.ai.model.State;
 
 public interface ActionSimulator {
     public State getNextState(State state, Action action);
