+++ src/main/java/com.ai/sim/Collision.java
@@ -1,4 +1,9 @@
-package com.ai;
+package com.ai.sim;
+
+import com.ai.Racetrack;
+import com.ai.model.Position;
+import com.ai.model.State;
+import com.ai.model.Velocity;
 
 public class Collision {
     public static final CollisionModel STOP = new StopCollisionModel();
@@ -85,10 +90,6 @@ public class Collision {
         }
         return (int)Math.floor(num);
     }
-}
-
-interface CollisionModel {
-    public State getNextState(Racetrack racetrack, Position position, Velocity velocity);
 }
 
 class StopCollisionModel implements CollisionModel {
